[FEAT] #75 fcm post 온보딩으로 넘김
