+++ app/src/main/java/com/timi/seulseul/presentation/onboarding/OnBoardingActivity.kt
@@ -1,17 +1,19 @@
 package com.timi.seulseul.presentation.onboarding
 
 import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import androidx.activity.viewModels
-import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
+import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
+import com.google.android.gms.tasks.OnCompleteListener
+import com.google.firebase.messaging.FirebaseMessaging
 import com.timi.seulseul.MainApplication.Companion.prefs
 import com.timi.seulseul.R
 import com.timi.seulseul.databinding.ActivityOnBoardingBinding
-import com.timi.seulseul.presentation.MainActivity
+import com.timi.seulseul.presentation.common.base.BaseActivity
 import com.timi.seulseul.presentation.permission.PermissionActivity
 import dagger.hilt.android.AndroidEntryPoint
+import timber.log.Timber
 
 @AndroidEntryPoint
 class OnBoardingActivity : AppCompatActivity() {
@@ -57,6 +59,9 @@ class OnBoardingActivity : AppCompatActivity() {
         // v1/user post uuid
         viewModel.postAuth()
 
+        // FCM 토큰 받기 & 보내기
+        getFcmToken()
+
     }
 
     private fun onConfirmButtonClick() {
@@ -68,5 +73,25 @@ class OnBoardingActivity : AppCompatActivity() {
 
         startActivity(Intent(this, PermissionActivity::class.java))
         finish()
+    }
+
+    private fun getFcmToken() {
+        FirebaseMessaging.getInstance().token.addOnCompleteListener(OnCompleteListener { task ->
+            if (!task.isSuccessful) {
+                Timber.d("Fetching FCM registeration token failed", task.exception)
+                return@OnCompleteListener
+            }
+
+            // 토큰값 가져오기
+            val token = task.result
+            BaseActivity.prefs.edit().putString("fcm_token", token).apply()
+            Timber.d("fcm_token: $token")
+
+            // 토큰값 보내기
+            viewModel.postFcmToken(token)
+
+        }).addOnFailureListener {
+            Timber.e(it)
+        }
     }
 }
